+++ wisdom/include/wisdom/dx12/impl/dx12_device.cpp
@@ -16,7 +16,14 @@ wis::DX12CreateDevice(wis::DX12AdapterHandle adapter) noexcept
     return wis::DX12CreateDeviceWithExtensions(adapter, nullptr, 0);
 }
 
-wis::ResultValue<wis::DX12Device> wis::DX12CreateDeviceWithExtensions(wis::DX12AdapterHandle adapter, wis::DX12DeviceExtension** extensions, uint32_t ext_count) noexcept
+wis::ResultValue<wis::DX12Device>
+wis::DX12CreateDeviceWithExtensions(wis::DX12AdapterHandle adapter, wis::DX12DeviceExtension** extensions, uint32_t ext_count) noexcept
+{
+    return wis::DX12CreateDeviceWithExtensionsForce(adapter, extensions, ext_count, false);
+}
+
+wis::ResultValue<wis::DX12Device>
+wis::DX12CreateDeviceWithExtensionsForce(wis::DX12AdapterHandle adapter, wis::DX12DeviceExtension** extensions, uint32_t ext_count, bool force) noexcept
 {
     auto in_adapter = std::get<0>(adapter);
 
@@ -32,6 +39,9 @@ wis::ResultValue<wis::DX12Device> wis::DX12CreateDeviceWithExtensions(wis::DX12A
         return wis::make_result<FUNC, "D3D12CreateDevice failed to create device">(hr);
 
     auto xdevice = wis::DX12Device(std::move(device), wis::com_ptr(in_adapter), std::move(in_factory));
+
+    if (!xdevice.QueryFeatureSupport(wis::DeviceFeature::EnchancedBarriers) && !force)
+        return wis::make_result<FUNC, "Device does not support enhanced barriers">(E_FAIL);
 
     for (uint32_t i = 0; i < ext_count; i++) {
         extensions[i]->Init(xdevice);
@@ -611,4 +621,25 @@ wis::DX12Device::CreateShaderResource(DX12TextureView texture, wis::ShaderResour
     device->CreateDescriptorHeap(&heap_desc, heap.iid(), heap.put_void());
     device->CreateShaderResourceView(std::get<0>(texture), &srv_desc, heap->GetCPUDescriptorHandleForHeapStart());
     return wis::DX12ShaderResource{ std::move(heap) };
+}
+
+bool wis::DX12Device::QueryFeatureSupport(wis::DeviceFeature feature) const noexcept
+{
+    switch (feature) {
+    case wis::DeviceFeature::EnchancedBarriers: {
+        D3D12_FEATURE_DATA_D3D12_OPTIONS12 options12 = {};
+        bool EnhancedBarriersSupported = false;
+        if (wis::succeeded(device->CheckFeatureSupport(D3D12_FEATURE_D3D12_OPTIONS12, &options12, sizeof(options12)))) {
+            EnhancedBarriersSupported = options12.EnhancedBarriersSupported;
+        }
+        return EnhancedBarriersSupported;
+    }
+    case wis::DeviceFeature::DescriptorBuffer:
+    case wis::DeviceFeature::WaitForPresent:
+    case wis::DeviceFeature::DescriptorEqualSize:
+    case wis::DeviceFeature::AdvancedIndexBuffer:
+        return true;
+    default:
+        return false;
+    }
 }